backend/requirements.txt
wheel==0.43.0
Django==4.2.13
django-cors-headers==4.4.0
django-dotenv==1.4.2
django-environ==0.11.2
djangorestframework==3.15.2
GitPython==3.1.41
jupyter==1.0.0
jupyter_client==8.6.0
jupyter-console==6.6.3
jupyter_core==5.5.0
jupyter-events==0.8.0
jupyter-lsp==2.2.0
jupyter_server==2.10.0
jupyter_server_fileid==0.9.0
jupyter-server-mathjax==0.2.6
jupyter_server_terminals==0.4.4
jupyter_server_ydoc==0.1.4
jupyterlab==4.0.11
jupyterlab_git==0.50.1
jupyterlab-pygments==0.1.2
jupyterlab_server==2.25.1
jupyterlab-widgets==3.0.10
nbdime==4.0.1
notebook==7.0.8
numba==0.58.1
numexpr==2.8.4
numpy==1.24.3
numpydoc==1.5.0
oauthlib==3.2.2
onnxruntime==1.17.1
openpyxl==3.1.2
pandas==2.0.3
passlib==1.7.4
pathlib==1.0.1
pillow==10.3.0
pip==24.2
poetry==1.8.3
poetry-core==1.9.0
poetry-plugin-export==1.8.0
portalocker==2.10.0
poyo==0.5.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
urllib3==1.26.18
virtualenv==20.26.2
vite==1.5.2
pyjwt